Improved offspring generation and multiprocessing
Crossover and mutations applied to lists of Genes. The multiprocessing is only used to create Individual objects from Gene lists which reduces the number of fitness evaluations. It also allows for cleaner multiprocessing code. The 'none' crossover option now utilises a force_mutation keyword that ensures that at leaste one gene is mutated in every offspring to avoid duplicate individuals in the combined population.